+++ hooks/post_gen_project.py
@@ -7,7 +7,7 @@ import sys
 try:
     from click.termui import secho
 except ImportError:
-    warn = note = print
+    warn = note = success = print
 else:
     def warn(text):
         for line in text.splitlines():
@@ -17,31 +17,13 @@ else:
         for line in text.splitlines():
             secho(line, fg="yellow", bold=True)
 
+    def success(text):
+        for line in text.splitlines():
+            secho(line, fg="green", bold=True)
+
+
 
 if __name__ == "__main__":
-{%- if cookiecutter.c_extension_test_pypi == 'yes' %}
-{%- if cookiecutter.test_matrix_separate_coverage == 'no' %}
-    warn("TODO: c_extension_test_pypi=yes will not work with test_matrix_separate_coverage=no for now.")
-    sys.exit(1)
-{%- endif %}
-{%- if cookiecutter.c_extension_support == 'no' %}
-    warn("""
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-!!                                                                            !!
-!!      ERROR:                                                                !!
-!!                                                                            !!
-!!          c_extension_test_pypi=yes is designed to build and publish        !!
-!!          platform-specific wheels.                                         !!
-!!                                                                            !!
-!!          You have set c_extension_support=no, and that will make every     !!
-!!          test environment publish duplicated universal wheels.             !!
-!!                                                                            !!
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-""")
-    sys.exit(1)
-{%- endif %}
-{%- endif %}
-
     cwd = pathlib.Path().resolve()
     src = cwd / 'src'
     ci = cwd / 'ci'
@@ -59,9 +41,8 @@ if __name__ == "__main__":
     src.joinpath('{{ cookiecutter.package_name }}', 'cli.py').unlink()
 {% endif %}
 
-{%- if cookiecutter.test_matrix_configurator == 'no' %}
-    ci.joinpath('templates', 'tox.ini').unlink()
-{% endif %}
+    ci.joinpath('templates', 'tox.ini').unlink(missing_ok=True)
+
 {%- if cookiecutter.allow_tests_inside_package == 'no' %}
     shutil.rmtree(src / '{{ cookiecutter.package_name }}' / 'tests')
 {% endif %}
@@ -83,20 +64,15 @@ if __name__ == "__main__":
     src.joinpath('{{ cookiecutter.package_name }}', '{{ cookiecutter.c_extension_module }}_build.py').unlink()
 {%- endif %}
 
-    ci.joinpath('appveyor-with-compiler.cmd').unlink(missing_ok=True)
-{%- if cookiecutter.appveyor == 'no' %}
-    ci.joinpath('templates', '.appveyor.yml').unlink()
-    cwd.joinpath('.appveyor.yml').unlink(missing_ok=True)
-{% endif %}
-    ci.joinpath('templates', 'appveyor.yml').unlink(missing_ok=True)
     ci.joinpath('appveyor-bootstrap.py').unlink(missing_ok=True)
     ci.joinpath('appveyor-download.py').unlink(missing_ok=True)
+    ci.joinpath('appveyor-with-compiler.cmd').unlink(missing_ok=True)
+    ci.joinpath('templates', '.appveyor.yml').unlink(missing_ok=True)
+    ci.joinpath('templates', 'appveyor.yml').unlink(missing_ok=True)
+    ci.joinpath('templates', '.travis.yml').unlink(missing_ok=True)
+    cwd.joinpath('.appveyor.yml').unlink(missing_ok=True)
     cwd.joinpath('appveyor.yml').unlink(missing_ok=True)
-
-{%- if cookiecutter.travis == 'no' %}
-    ci.joinpath('templates', '.travis.yml').unlink()
     cwd.joinpath('.travis.yml').unlink(missing_ok=True)
-{% endif %}
 
 {%- if cookiecutter.github_actions == 'no' %}
     ci.joinpath('templates', '.github', 'workflows', 'github-actions.yml').unlink()
@@ -113,10 +89,6 @@ if __name__ == "__main__":
     src.joinpath('{{ cookiecutter.package_name }}', '_version.py').unlink(missing_ok=True)
 {% endif %}
 
-{%- if cookiecutter.pre_commit == 'no' %}
-    cwd.joinpath('.pre-commit-config.yaml').unlink()
-{% endif %}
-
 {%- if cookiecutter.version_manager == 'bump2version' %}
     cwd.joinpath('tbump.toml').unlink()
 {%- elif cookiecutter.version_manager == 'tbump' %}
@@ -126,12 +98,10 @@ if __name__ == "__main__":
 {%- if cookiecutter.license == "no" %}
     cwd.joinpath('LICENSE').unlink()
 {% endif %}
+    cwd.joinpath('setup.cfg').unlink(missing_ok=True)
 
-    print("""
-################################################################################
-
-    Generating CI configuration ...
-""")
+    width = min(140, shutil.get_terminal_size(fallback=(140, 0)).columns)
+    note(" Generating CI configuration ".center(width, "#"))
     try:
         subprocess.check_call(['tox', '-e', 'bootstrap', '--sitepackages'])
     except Exception:
@@ -139,88 +109,50 @@ if __name__ == "__main__":
             subprocess.check_call([sys.executable, '-mtox', '-e', 'bootstrap', '--sitepackages'])
         except Exception:
             subprocess.check_call([sys.executable, ci / 'bootstrap.py'])
-
-    print("""
-################################################################################
-################################################################################
-
-    You have succesfully created `{{ cookiecutter.repo_name }}`.
-
-################################################################################
-
-    You've used these cookiecutter parameters:
-{% for key, value in cookiecutter.items()|sort %}
-        {{ "{0:26}".format(key + ":") }} {{ "{0!r}".format(value).strip("u") }}
-{%- endfor %}
-
-    See .cookiecutterrc for instructions on regenerating the project.
-
-################################################################################
-
-    To get started run these:
-
-        cd {{ cookiecutter.repo_name }}
-        git init
-        git add --all
-        git commit -m "Add initial project skeleton."
-        git tag v{{ cookiecutter.version }}
-        git remote add origin git@{{ cookiecutter.repo_hosting_domain }}:{{ cookiecutter.repo_username }}/{{ cookiecutter.repo_name }}.git
-        git push -u origin {{ cookiecutter.repo_main_branch }} v{{ cookiecutter.version }}
-
-{% if cookiecutter.test_matrix_configurator == "yes" %}
-    To regenerate your tox.ini, .travis.yml or .appveyor.yml run:
-{% else %}
-    To regenerate your .travis.yml or .appveyor.yml run:
-{% endif %}
-
-        tox -e bootstrap
-
-    You can also run:
-
-        ./ci/bootstrap.py
-
-""")
-
-{%- if cookiecutter.c_extension_test_pypi == 'yes' %}
-    note("""
-NOTE:
-
-    You are using the c_extension_test_pypi option.
-
-    Make sure you are setting TWINE_PASSWORD as a secret env variable in CI settings:
-    - https://ci.appveyor.com/project/{{ cookiecutter.repo_username }}/{{ cookiecutter.repo_name }}/settings/environment
-    - https://travis-ci.com{% if cookiecutter.repo_hosting == 'github.com' %}/github
-                           {%- elif cookiecutter.repo_hosting == 'gitlab.com' %}/gitlab
-                           {%- endif %}/{{ cookiecutter.repo_username }}/{{ cookiecutter.repo_name }}/settings
-""")
-{%- endif %}
-
+    note(' Setting up pre-commit '.center(width, "#"))
+    if cwd.joinpath('.git').exists():
+        subprocess.check_call(['pre-commit', 'install', '--install-hooks'])
+        subprocess.check_call(['pre-commit', 'autoupdate'])
+    else:
+        print('Skipping precommit install.')
+    success(' Successfully created `{{ cookiecutter.repo_name }}` '.center(width, "#"))
+    print('See .cookiecutterrc for instructions on regenerating the project.')
+    note('To get started run these:')
+    print('''
+cd {{ cookiecutter.repo_name }}
+git init
+pre-commit install --install-hooks
+pre-commit autoupdate
+git add --all
+git commit -m "Add initial project skeleton."
+git tag v{{ cookiecutter.version }}
+git remote add origin git@{{ cookiecutter.repo_hosting_domain }}:{{ cookiecutter.repo_username }}/{{ cookiecutter.repo_name }}.git
+git push -u origin {{ cookiecutter.repo_main_branch }} v{{ cookiecutter.version }}
+''')
     command_line_interface_bin_name = '{{ cookiecutter.command_line_interface_bin_name }}'
     while command_line_interface_bin_name.endswith('.py'):
         command_line_interface_bin_name = command_line_interface_bin_name[:-3]
 
         if command_line_interface_bin_name == '{{ cookiecutter.package_name }}':
-            warn("""
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-!!                                                                            !!
-!!      ERROR:                                                                !!
-!!                                                                            !!
-!!          Your result package is broken. Your bin script named              !!
-!!          {0} !!
-!!                                                                            !!
-!!          Python automatically adds the location of scripts to              !!
-!!          `sys.path`. Because of that, the bin script will fail             !!
-!!          to import your package because it has the same name               !!
-!!          (it will try to import itself as a module).                       !!
-!!                                                                            !!
-!!          To avoid this problem you have two options:                       !!
-!!                                                                            !!
-!!          * Remove the ".py" suffix from the `command_line_interface_bin_name`.                    !!
-!!                                                                            !!
-!!          * Use a different `package_name` {1} !!
-!!                                                                            !!
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-""".format(
+            warn('''
+┌───────────────────────────────────────────────────────────────────────┐
+│ ERROR:                                                                │
+│                                                                       │
+│     Your result package is broken. Your bin script named              │
+│     {0} │
+│                                                                       │
+│     Python automatically adds the location of scripts to              │
+│     `sys.path`. Because of that, the bin script will fail             │
+│     to import your package because it has the same name               │
+│     (it will try to import itself as a module).                       │
+│                                                                       │
+│     To avoid this problem you have two options:                       │
+│                                                                       │
+│     * Remove the ".py" suffix from `command_line_interface_bin_name`. │
+│                                                                       │
+│     * Use a different `package_name` {1} │
+└───────────────────────────────────────────────────────────────────────┘
+'''.format(
                 '"{{ cookiecutter.command_line_interface_bin_name }}" will shadow your package.'.ljust(65),
                 '(not "{0}").'.format(command_line_interface_bin_name).ljust(32)))
             sys.exit(1)