+++ {{cookiecutter.repo_name}}/ci/bootstrap.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 import os
 import pathlib
 import subprocess
@@ -23,7 +22,7 @@ def exec_in_env():
     if not env_path.exists():
         import subprocess
 
-        print("Making bootstrap env in: {0} ...".format(env_path))
+        print(f"Making bootstrap env in: {env_path} ...")
         try:
             check_call([sys.executable, "-m", "venv", env_path])
         except subprocess.CalledProcessError:
@@ -32,23 +31,20 @@ def exec_in_env():
             except subprocess.CalledProcessError:
                 check_call(["virtualenv", env_path])
         print("Installing `jinja2` into bootstrap environment...")
-        check_call([bin_path / "pip", "install", "jinja2", "tox"{% if cookiecutter.test_matrix_configurator == "yes" %}, "matrix"{% endif %}])
+        check_call([bin_path / "pip", "install", "jinja2", "tox"])
     python_executable = bin_path / "python"
     if not python_executable.exists():
-        python_executable = python_executable.with_suffix('.exe')
+        python_executable = python_executable.with_suffix(".exe")
 
-    print("Re-executing with: {0}".format(python_executable))
+    print(f"Re-executing with: {python_executable}")
     print("+ exec", python_executable, __file__, "--no-env")
     os.execv(python_executable, [python_executable, __file__, "--no-env"])
 
 
 def main():
     import jinja2
-{%- if cookiecutter.test_matrix_configurator == "yes" %}
-    import matrix
-{%- endif %}
 
-    print("Project path: {0}".format(base_path))
+    print(f"Project path: {base_path}")
 
     jinja = jinja2.Environment(
         loader=jinja2.FileSystemLoader(str(templates_path)),
@@ -56,20 +52,6 @@ def main():
         lstrip_blocks=True,
         keep_trailing_newline=True,
     )
-{% if cookiecutter.test_matrix_configurator == "yes" %}
-    tox_environments = {}
-    for (alias, conf) in matrix.from_file(base_path / "setup.cfg").items():
-        deps = conf["dependencies"]
-        tox_environments[alias] = {
-            "deps": deps.split(),
-        }
-        if "coverage_flags" in conf:
-            cover = {"false": False, "true": True}[conf["coverage_flags"].lower()]
-            tox_environments[alias].update(cover=cover)
-        if "environment_variables" in conf:
-            env_vars = conf["environment_variables"]
-            tox_environments[alias].update(env_vars=env_vars.split())
-{% else %}
     tox_environments = [
         line.strip()
         # 'tox' need not be installed globally, but must be importable
@@ -77,17 +59,16 @@ def main():
         # This uses sys.executable the same way that the call in
         # cookiecutter-pylibrary/hooks/post_gen_project.py
         # invokes this bootstrap.py itself.
-        for line in subprocess.check_output([sys.executable, '-m', 'tox', '--listenvs'], universal_newlines=True).splitlines()
+        for line in subprocess.check_output([sys.executable, "-m", "tox", "--listenvs"], universal_newlines=True).splitlines()
     ]
-    tox_environments = [line for line in tox_environments if line.startswith('py')]
-{% endif %}
-    for template in templates_path.rglob('*'):
+    tox_environments = [line for line in tox_environments if line.startswith("py")]
+    for template in templates_path.rglob("*"):
         if template.is_file():
             template_path = str(template.relative_to(templates_path))
             destination = base_path / template_path
             destination.parent.mkdir(parents=True, exist_ok=True)
             destination.write_text(jinja.get_template(template_path).render(tox_environments=tox_environments))
-            print("Wrote {}".format(template_path))
+            print(f"Wrote {template_path}")
     print("DONE.")
 
 
@@ -98,5 +79,5 @@ if __name__ == "__main__":
     elif not args:
         exec_in_env()
     else:
-        print("Unexpected arguments {0}".format(args), file=sys.stderr)
+        print(f"Unexpected arguments: {args}", file=sys.stderr)
         sys.exit(1)