+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.package_name}}/__init__.py
@@ -1,25 +1,30 @@
-__version__ = '{{ cookiecutter.version }}'
-{%- if cookiecutter.c_extension_support == 'cffi' %}
+__version__ = "{{ cookiecutter.version }}"
+{%- if cookiecutter.c_extension_support == "cffi" %}
 
 from .{{ cookiecutter.c_extension_module }} import ffi as _ffi
 from .{{ cookiecutter.c_extension_module }} import lib as _lib
 
 
 def {{ cookiecutter.c_extension_function }}(args):
-    args = [_ffi.new('char[]', arg) for arg in args]
-    result = _lib.{{ cookiecutter.c_extension_function }}(len(args), _ffi.new('char *[]', args))
+    args = [_ffi.new("char[]", arg) for arg in args]
+    result = _lib.{{ cookiecutter.c_extension_function }}(len(args), _ffi.new("char *[]", args))
     if result == _ffi.NULL:
-        return ''
+        return ""
     else:
         return _ffi.string(result)
-{%- elif cookiecutter.c_extension_support != 'no' %}
-{%- if cookiecutter.c_extension_optional == 'yes' %}
+{%- elif cookiecutter.c_extension_support != "no" %}
+{%- if cookiecutter.c_extension_optional == "yes" %}
 try:
-    from .{{ cookiecutter.c_extension_module }} import {{ cookiecutter.c_extension_function }}  # noqa
+    from .{{ cookiecutter.c_extension_module }} import {{ cookiecutter.c_extension_function }}
 except ImportError:
+
     def {{ cookiecutter.c_extension_function }}(args):
         return max(args, key=len)
 {%- else %}
-from .{{ cookiecutter.c_extension_module }} import {{ cookiecutter.c_extension_function }}  # noqa
+from .{{ cookiecutter.c_extension_module }} import {{ cookiecutter.c_extension_function }}
+
+__all__ = [
+    "{{ cookiecutter.c_extension_function }}",
+]
 {%- endif %}
 {%- endif %}