+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.package_name}}/__main__.py
@@ -8,13 +8,13 @@ Why does this file exist, and why __main__? For more info, read:
 - https://docs.python.org/2/using/cmdline.html#cmdoption-m
 - https://docs.python.org/3/using/cmdline.html#cmdoption-m
 """
-{%- if cookiecutter.command_line_interface == 'plain' %}
+{%- if cookiecutter.command_line_interface == "plain" %}
 import sys
 {% endif %}
 from {{cookiecutter.package_name}}.cli import main
 
 if __name__ == "__main__":
-{%- if cookiecutter.command_line_interface == 'plain' %}
+{%- if cookiecutter.command_line_interface == "plain" %}
     sys.exit(main())
 {%- else %}
     main()