+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.package_name}}/cli.py
@@ -7,32 +7,36 @@ Why does this file exist, and why not put this in __main__?
   problems: the code will get executed twice:
 
   - When you run `python -m{{cookiecutter.package_name}}` python will execute
-    ``__main__.py`` as a script. That means there won't be any
+    ``__main__.py`` as a script. That means there will not be any
     ``{{cookiecutter.package_name}}.__main__`` in ``sys.modules``.
   - When you import __main__ it will get executed again (as a module) because
-    there's no ``{{cookiecutter.package_name}}.__main__`` in ``sys.modules``.
+    there"s no ``{{cookiecutter.package_name}}.__main__`` in ``sys.modules``.
 
   Also see (1) from http://click.pocoo.org/5/setuptools/#setuptools-integration
 """
-{%- if cookiecutter.command_line_interface == 'click' %}
+{%- if cookiecutter.command_line_interface == "click" %}
 import click
-{%- elif cookiecutter.command_line_interface == 'argparse' %}
+{%- elif cookiecutter.command_line_interface == "argparse" %}
 import argparse
 {%- else %}
 import sys
 {%- endif %}
-{%- if cookiecutter.command_line_interface == 'click' %}
+{%- if cookiecutter.command_line_interface == "click" %}
 
 
 @click.command()
-@click.argument('names', nargs=-1)
+@click.argument("names", nargs=-1)
 def main(names):
     click.echo(repr(names))
-{%- elif cookiecutter.command_line_interface == 'argparse' %}
+{%- elif cookiecutter.command_line_interface == "argparse" %}
 
-parser = argparse.ArgumentParser(description='Command description.')
-parser.add_argument('names', metavar='NAME', nargs=argparse.ZERO_OR_MORE,
-                    help="A name of something.")
+parser = argparse.ArgumentParser(description="Command description.")
+parser.add_argument(
+    "names",
+    metavar="NAME",
+    nargs=argparse.ZERO_OR_MORE,
+    help="A name of something.",
+)
 
 
 def main(args=None):