+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.package_name}}/{{cookiecutter.c_extension_module}}_build.py
@@ -1,17 +1,19 @@
-from os.path import dirname
-from os.path import join
+from pathlib import Path
 
 from cffi import FFI
 
 ffi = FFI()
-ffi.cdef('''
+ffi.cdef(
+    """
     char* {{ cookiecutter.c_extension_function }}(int argv, char *argv[]);
-''')
-
-ffi.set_source(
-    '{{ cookiecutter.package_name }}.{{ cookiecutter.c_extension_module }}',
-    open(join(dirname(__file__), '{{ cookiecutter.c_extension_module }}.c')).read()
+    """
 )
 
-if __name__ == '__main__':
+
+ffi.set_source(
+    "{{ cookiecutter.package_name }}.{{ cookiecutter.c_extension_module }}",
+    Path(__file__).parent.joinpath("{{ cookiecutter.c_extension_module }}.c").read_text(),
+)
+
+if __name__ == "__main__":
     ffi.compile()